refactor:station结构体

+++ model/station.go
@@ -3,8 +3,8 @@ package model
 type Station struct {
 	Id           int
 	Name         string
-	Longitude    float32
-	Latitude     float32
+	Longitude    float64
+	Latitude     float64
 	One          int
 	Two          int
 	Three        int
